Added Index.html & Styes.css
HTML is compliant with Assignment's required HTML Structure.  
HTML is screen-size responsive.
Home Content is included. 
There is no LINKS link. (Should be added, I'd like it was a separate box which sits under the nav menu, and goes to the end of page on @mediaQuery)

## Extra additons
1. There is also a GMAPS iframe at the bottom (it is a no API gmap, just for showcasing what i'd look like if we had one)
2. There is a search query box (not functional, again just for showcasing what i'd look like if we had one)

There is no actual stylization.